+++ src/repr/edge_list/mod.rs
@@ -120,8 +120,11 @@ use {
         ErdosRenyi,
         GrowingNetwork,
         HasArc,
+        InNeighbors,
         Indegree,
+        IndegreeSequence,
         IsComplete,
+        IsRegular,
         IsSemicomplete,
         IsSimple,
         IsTournament,
@@ -129,6 +132,7 @@ use {
         OutNeighbors,
         OutNeighborsWeighted,
         Outdegree,
+        OutdegreeSequence,
         Path,
         RandomTournament,
         RemoveArc,
@@ -548,9 +552,21 @@ impl Indegree for EdgeList {
     }
 }
 
+impl InNeighbors for EdgeList {
+    fn in_neighbors(&self, v: usize) -> impl Iterator<Item = usize> {
+        self.arcs().filter_map(move |(x, y)| (v == y).then_some(x))
+    }
+}
+
 impl IsComplete for EdgeList {
     fn is_complete(&self) -> bool {
         *self == Self::complete(self.order())
+    }
+}
+
+impl IsRegular for EdgeList {
+    fn is_regular(&self) -> bool {
+        self.indegree_sequence().eq(self.outdegree_sequence())
     }
 }